[refactor] 코드 최적화 - 페이지 랜더링 - 카테고리 기능
