--- src/pages/Communities/Community/Community.css
@@ -1,127 +1,0 @@
-.community_main {
-  width: 100%;
-  margin: auto;
-  margin-top: 190px;
-  padding: 20px;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-/* 헤더 */
-
-.community_header {
-  width: 100%;
-
-  font-family: "NanumSquareNeo", sans-serif;
-  font-size: 42px;
-  font-weight: 700;
-  
-  margin-bottom: 30px;
-}
-.community_header div {
-  margin-left: 85px;
-}
-.community_header span {
-  font-weight: 400;
-  font-size: 30px;
-}
-
-.community_search {
-  width: 80%;
-
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-
-  gap: 10px;
-  margin-bottom: 20px;
-}
-  
-  .category_select, 
-  input[type="text"], 
-  .search_button {
-    padding: 8px;
-    font-size: 16px;
-  }
-  
-  .search_button {
-    background-color: #1a73e8;
-    color: white;
-    border: none;
-    cursor: pointer;
-  }
-
-  .no_results {
-    text-align: center;
-    padding: 20px;
-    font-size: 18px;
-    color: #666;
-  }
-  
-
-  /* 테이블 */
-  
-  .community_table {
-    width: 80%;
-    border-collapse: collapse;
-    text-align: center;
-    font-weight: 500;
-
-  }
-  
-  .community_table th, .community_table td {
-    padding: 10px;
-    vertical-align: middle; /* 세로 중앙 정렬 */
-    height: 60px;
-    
-    border-top: 1px solid #ddd;
-    border-bottom: 1px solid #ddd;
-
-  }
-  
-  .community_table thead {
-    background-color: var(--color-navy);
-    color: var(--color-white);
-    height: 80px;
-    font-weight: 700;
-    font-size: 20px;
-  }
-
-  .community_pointer{
-    cursor: pointer;
-  }
-
-  .community_table tbody tr:nth-child(even) {
-    background-color: var(--color-main);
-  }
-  
-  .community_table tbody tr:hover{
-    background-color: #b2c0d3;
-  }
-
-
-  /* 페이지네이션 */
-.community_pagination {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 25px;
-  margin-top: 20px;
-}
-
-.community_pagination a {
-  text-decoration: none;
-  color: #333;
-  font-size: 20px;
-}
-
-.community_pagination a:hover {
-  font-weight: bold;
-}
-
-/* 현재 페이지 강조 */
-.community_pagination a.active {
-  font-weight: bold;
-}
