+++ src/pages/Communities/Community/dataCommunity.css
@@ -9,40 +9,6 @@
   align-items: center;
 }
 
-.community_category {
-  height: 45px;
-  width: 80%;
-
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.category_left {
-  display: flex;
-  align-items: center;
-  gap: 35px;
-  height: 100%;
-}
-
-.category_bg {
-  height: 100%;
-  padding: 0 12px;
-  background-color: #6e83a2;
-  color: var(--color-white);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-/* 선택된 게시판 스타일 */
-.category_bg.active {
-  background-color: var(--color-navy);
-  font-weight: bold;
-}
-
 /* 헤더 */
 
 .community_header {
@@ -62,6 +28,38 @@
   font-size: 30px;
 }
 
+.community_search {
+  width: 80%;
+
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+
+  gap: 10px;
+  margin-bottom: 20px;
+}
+
+.category_select,
+input[type="text"],
+.search_button {
+  padding: 8px;
+  font-size: 16px;
+}
+
+.search_button {
+  background-color: #1a73e8;
+  color: white;
+  border: none;
+  cursor: pointer;
+}
+
+.no_results {
+  text-align: center;
+  padding: 20px;
+  font-size: 18px;
+  color: #666;
+}
+
 /* 테이블 */
 
 .community_table {
@@ -76,6 +74,9 @@
   padding: 10px;
   vertical-align: middle; /* 세로 중앙 정렬 */
   height: 60px;
+
+  border-top: 1px solid #ddd;
+  border-bottom: 1px solid #ddd;
 }
 
 .community_table thead {
@@ -103,8 +104,8 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  gap: 30px;
-  margin: 20px;
+  gap: 25px;
+  margin-top: 20px;
 }
 
 .community_pagination a {
@@ -120,36 +121,4 @@
 /* 현재 페이지 강조 */
 .community_pagination a.active {
   font-weight: bold;
-}
-
-.open_community_search {
-  width: 80%;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  gap: 10px;
-  margin-bottom: 20px;
-}
-
-.category_select,
-input[type="text"],
-.search_button {
-  padding: 8px;
-  font-size: 16px;
-}
-
-.search_button {
-  background-color: #1a73e8;
-  color: white;
-  border: none;
-  cursor: pointer;
-}
-
-.no_results {
-  text-align: center;
-  padding: 20px;
-  font-size: 18px;
-  color: #666;
 }
